cloud-functions-v1/requirements.txt
functions-framework==3.*
google-api-python-client
google-auth
apache-beam[gcp]